Majd_requirements.txt
fitz==0.0.1.dev2
openpyxl==3.1.5
pdf2image==1.17.0
pdfplumber==0.11.5
Pillow==11.2.1
pytesseract==0.3.13
python_docx==1.1.2
streamlit==1.41.1
tqdm==4.67.1